requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
newspaper3k==0.2.8
lxml_html_clean  # нужно для newspaper3k